moved api functions to avoid CORS rejection
